reorganize util package
