+++ catalog-hive/src/main/java/com/datastrato/graviton/catalog/hive/HiveClientPool.java
@@ -19,7 +19,7 @@
 package com.datastrato.graviton.catalog.hive;
 
 import com.datastrato.graviton.catalog.hive.dyn.DynMethods;
-import com.datastrato.graviton.util.ClientPoolImpl;
+import com.datastrato.graviton.utils.ClientPoolImpl;
 import com.google.common.annotations.VisibleForTesting;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.conf.HiveConf;
