+++ core/src/main/java/com/datastrato/graviton/catalog/CatalogManager.java
@@ -24,8 +24,8 @@ import com.datastrato.graviton.meta.BaseMetalake;
 import com.datastrato.graviton.meta.CatalogEntity;
 import com.datastrato.graviton.rel.SupportsSchemas;
 import com.datastrato.graviton.rel.TableCatalog;
-import com.datastrato.graviton.util.IsolatedClassLoader;
-import com.datastrato.graviton.util.ThrowableFunction;
+import com.datastrato.graviton.utils.IsolatedClassLoader;
+import com.datastrato.graviton.utils.ThrowableFunction;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.github.benmanes.caffeine.cache.Scheduler;
