+++ core/src/main/java/com/datastrato/graviton/storage/kv/BinaryEntityKeyEncoder.java
@@ -13,8 +13,8 @@ import com.datastrato.graviton.Entity.EntityType;
 import com.datastrato.graviton.NameIdentifier;
 import com.datastrato.graviton.storage.EntityKeyEncoder;
 import com.datastrato.graviton.storage.NameMappingService;
-import com.datastrato.graviton.util.ByteUtils;
-import com.datastrato.graviton.util.Bytes;
+import com.datastrato.graviton.utils.ByteUtils;
+import com.datastrato.graviton.utils.Bytes;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
