+++ core/src/main/java/com/datastrato/graviton/storage/kv/KvEntityStore.java
@@ -19,8 +19,8 @@ import com.datastrato.graviton.Namespace;
 import com.datastrato.graviton.exceptions.NoSuchEntityException;
 import com.datastrato.graviton.storage.EntityKeyEncoder;
 import com.datastrato.graviton.storage.NameMappingService;
-import com.datastrato.graviton.util.Bytes;
-import com.datastrato.graviton.util.Executable;
+import com.datastrato.graviton.utils.Bytes;
+import com.datastrato.graviton.utils.Executable;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
