+++ core/src/main/java/com/datastrato/graviton/storage/kv/KvNameMappingService.java
@@ -8,8 +8,8 @@ package com.datastrato.graviton.storage.kv;
 import com.datastrato.graviton.storage.IdGenerator;
 import com.datastrato.graviton.storage.NameMappingService;
 import com.datastrato.graviton.storage.RandomIdGenerator;
-import com.datastrato.graviton.util.ByteUtils;
-import com.datastrato.graviton.util.Bytes;
+import com.datastrato.graviton.utils.ByteUtils;
+import com.datastrato.graviton.utils.Bytes;
 import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
