+++ core/src/main/java/com/datastrato/graviton/storage/kv/RocksDBKvBackend.java
@@ -8,9 +8,9 @@ package com.datastrato.graviton.storage.kv;
 import com.datastrato.graviton.Config;
 import com.datastrato.graviton.Configs;
 import com.datastrato.graviton.EntityAlreadyExistsException;
-import com.datastrato.graviton.util.ByteUtils;
-import com.datastrato.graviton.util.Bytes;
-import com.datastrato.graviton.util.Executable;
+import com.datastrato.graviton.utils.ByteUtils;
+import com.datastrato.graviton.utils.Bytes;
+import com.datastrato.graviton.utils.Executable;
 import com.google.common.base.Throwables;
 import com.google.common.collect.Lists;
 import java.io.File;
