+++ core/src/main/java/com/datastrato/graviton/utils/ThrowableFunction.java
@@ -2,10 +2,10 @@
  * Copyright 2023 Datastrato.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.graviton.util;
+package com.datastrato.graviton.utils;
 
 @FunctionalInterface
-public interface Executable<R, E extends Exception> {
+public interface ThrowableFunction<T, R> {
 
-  R execute() throws E;
+  R apply(T t) throws Exception;
 }
