+++ core/src/test/java/com/datastrato/graviton/TestField.java
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
-class FieldTest {
+class TestField {
 
   @Test
   void testRequiredFieldCreation() {
