+++ core/src/test/java/com/datastrato/graviton/storage/kv/TestEntityKeyEncoding.java
@@ -22,8 +22,8 @@ import com.datastrato.graviton.NameIdentifier;
 import com.datastrato.graviton.Namespace;
 import com.datastrato.graviton.storage.IdGenerator;
 import com.datastrato.graviton.storage.NameMappingService;
-import com.datastrato.graviton.util.ByteUtils;
-import com.datastrato.graviton.util.Bytes;
+import com.datastrato.graviton.utils.ByteUtils;
+import com.datastrato.graviton.utils.Bytes;
 import java.io.IOException;
 import java.lang.reflect.Field;
 import org.apache.commons.io.FileUtils;
