+++ core/src/test/java/com/datastrato/graviton/utils/TestBytes.java
@@ -9,7 +9,6 @@ import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.datastrato.graviton.util.Bytes;
 import org.junit.jupiter.api.Test;
 
 public class TestBytes {
